// src/interfaces.ts
import { VERSION_NUMBER, CREATED_PROPERTIES, DELETED_PROPERTIES, MODIFIED_VALUES } from './types'

export interface HistoryObject {
	VERSION_NUMBER: VERSION_NUMBER
	CREATED_PROPERTIES: CREATED_PROPERTIES
	DELETED_PROPERTIES: DELETED_PROPERTIES
	MODIFIED_VALUES: MODIFIED_VALUES
}

export interface Options {
	considerVersionNumbers?: boolean
}